OGE 1.0
fully fixed all the bugs, improved user interface and made it actually usable and fairly obvious to use

+++ Src/Headers/Utils/VideoMemory/VertexBufferHandle.hpp
@@ -2,5 +2,5 @@
 #include "Calculations/General/Types.hpp"
 class VertexBufferHandle {
 public:
-	u32 index = {};
+	usize index = {};
 };
